test(gdn): add torch reference for Triton prefill
The existing GDN prefill test compared the public Triton registry path against the same FLA chunk helper that the registry calls internally. That is useful for coverage of selection and argument plumbing, but it does not independently validate the recurrence math.

Add a small PyTorch token-by-token reference for the GDN recurrence and use it to check the Triton gdn_chunk_prefill path. The test uses Qwen3-style prefill dimensions for the single-sequence case and a smaller varlen case for cu_seqlens coverage, both with qk_l2norm=True so the random inputs stay in the stable production-like regime.

Keep FlashInfer covered by the existing FLA parity test, since the torch reference test is intended to be a portable semantic check for the Triton implementation rather than a Blackwell-specific backend test.
