+++ tokenspeed-kernel/python/tokenspeed_kernel/ops/attention/triton/linear/chunk_o.py
@@ -27,8 +27,8 @@
 
 import torch
 from tokenspeed_kernel._triton import tl, triton
-from tokenspeed_kernel.ops.attention.triton.fla.index import prepare_chunk_indices
-from tokenspeed_kernel.ops.attention.triton.fla.op import exp, safe_exp
+from tokenspeed_kernel.ops.attention.triton.linear.index import prepare_chunk_indices
+from tokenspeed_kernel.ops.attention.triton.linear.op import exp, safe_exp
 
 
 @triton.heuristics(
