+++ tokenspeed-kernel/python/tokenspeed_kernel/ops/attention/triton/linear/cumsum.py
@@ -27,8 +27,8 @@
 
 import torch
 from tokenspeed_kernel._triton import tl, triton
-from tokenspeed_kernel.ops.attention.triton.fla.index import prepare_chunk_indices
-from tokenspeed_kernel.ops.attention.triton.fla.utils import (
+from tokenspeed_kernel.ops.attention.triton.linear.index import prepare_chunk_indices
+from tokenspeed_kernel.ops.attention.triton.linear.utils import (
     check_shared_mem,
     input_guard,
 )
