+++ tokenspeed-kernel/python/tokenspeed_kernel/ops/attention/triton/linear/l2norm.py
@@ -27,7 +27,7 @@
 
 import torch
 from tokenspeed_kernel._triton import tl, triton
-from tokenspeed_kernel.ops.attention.triton.fla.utils import input_guard
+from tokenspeed_kernel.ops.attention.triton.linear.utils import input_guard
 
 
 @triton.jit
