+++ tokenspeed-kernel/python/tokenspeed_kernel/ops/attention/triton/linear/wy_fast.py
@@ -27,32 +27,30 @@
 
 import torch
 from tokenspeed_kernel._triton import tl, triton
-from tokenspeed_kernel.ops.attention.triton.fla.index import prepare_chunk_indices
-from tokenspeed_kernel.ops.attention.triton.fla.op import safe_exp
+from tokenspeed_kernel.ops.attention.triton.linear.index import prepare_chunk_indices
 
 
-@triton.heuristics(
-    {
-        "IS_VARLEN": lambda args: args["cu_seqlens"] is not None,
-        "USE_G": lambda args: args["g_cumsum"] is not None,
-    }
-)
+@triton.heuristics({"IS_VARLEN": lambda args: args["cu_seqlens"] is not None})
 @triton.jit(do_not_specialize=["T"])
-def chunk_scaled_dot_kkt_fwd_kernel(
+def recompute_w_u_fwd_kernel(
     k,
+    v,
     beta,
-    g_cumsum,
+    w,
+    u,
     A,
+    g,
     cu_seqlens,
     chunk_indices,
     T,
     H: tl.constexpr,
     Hg: tl.constexpr,
     K: tl.constexpr,
+    V: tl.constexpr,
     BT: tl.constexpr,
     BK: tl.constexpr,
+    BV: tl.constexpr,
     IS_VARLEN: tl.constexpr,
-    USE_G: tl.constexpr,
 ):
     i_t, i_bh = tl.program_id(0), tl.program_id(1)
     i_b, i_h = i_bh // H, i_bh % H
@@ -66,14 +64,39 @@ def chunk_scaled_dot_kkt_fwd_kernel(
         T = eos - bos
     else:
         bos, eos = i_b * T, i_b * T + T
-    o_t = tl.arange(0, BT)
-
     p_beta = tl.make_block_ptr(
         beta + bos * H + i_h, (T,), (H,), (i_t * BT,), (BT,), (0,)
     )
+    p_g = tl.make_block_ptr(g + (bos * H + i_h), (T,), (H,), (i_t * BT,), (BT,), (0,))
+    p_A = tl.make_block_ptr(
+        A + (bos * H + i_h) * BT, (T, BT), (H * BT, 1), (i_t * BT, 0), (BT, BT), (1, 0)
+    )
     b_beta = tl.load(p_beta, boundary_check=(0,))
+    b_A = tl.load(p_A, boundary_check=(0, 1))
+    b_g = tl.exp(tl.load(p_g, boundary_check=(0,)))
 
-    b_A = tl.zeros([BT, BT], dtype=tl.float32)
+    for i_v in range(tl.cdiv(V, BV)):
+        p_v = tl.make_block_ptr(
+            v + (bos * H + i_h) * V,
+            (T, V),
+            (H * V, 1),
+            (i_t * BT, i_v * BV),
+            (BT, BV),
+            (1, 0),
+        )
+        p_u = tl.make_block_ptr(
+            u + (bos * H + i_h) * V,
+            (T, V),
+            (H * V, 1),
+            (i_t * BT, i_v * BV),
+            (BT, BV),
+            (1, 0),
+        )
+        b_v = tl.load(p_v, boundary_check=(0, 1))
+        b_vb = (b_v * b_beta[:, None]).to(b_v.dtype)
+        b_u = tl.dot(b_A, b_vb, allow_tf32=False)
+        tl.store(p_u, b_u.to(p_u.dtype.element_ty), boundary_check=(0, 1))
+
     for i_k in range(tl.cdiv(K, BK)):
         p_k = tl.make_block_ptr(
             k + (bos * Hg + i_h // (H // Hg)) * K,
@@ -83,79 +106,62 @@ def chunk_scaled_dot_kkt_fwd_kernel(
             (BT, BK),
             (1, 0),
         )
-        b_k = tl.load(p_k, boundary_check=(0, 1))
-        b_kb = b_k * b_beta[:, None]
-        b_A += tl.dot(b_kb.to(b_k.dtype), tl.trans(b_k))
-
-    if USE_G:
-        p_g = tl.make_block_ptr(
-            g_cumsum + bos * H + i_h, (T,), (H,), (i_t * BT,), (BT,), (0,)
+        p_w = tl.make_block_ptr(
+            w + (bos * H + i_h) * K,
+            (T, K),
+            (H * K, 1),
+            (i_t * BT, i_k * BK),
+            (BT, BK),
+            (1, 0),
         )
-        b_g = tl.load(p_g, boundary_check=(0,))
-        b_g_diff = b_g[:, None] - b_g[None, :]
-        b_A = b_A * safe_exp(b_g_diff)
-
-    b_A = tl.where(o_t[:, None] > o_t[None, :], b_A, 0)
-    p_A = tl.make_block_ptr(
-        A + (bos * H + i_h) * BT, (T, BT), (BT * H, 1), (i_t * BT, 0), (BT, BT), (1, 0)
-    )
-    tl.store(p_A, b_A.to(p_A.dtype.element_ty), boundary_check=(0, 1))
+        b_k = tl.load(p_k, boundary_check=(0, 1))
+        b_kb = (b_k * b_beta[:, None] * b_g[:, None]).to(b_k.dtype)
+        b_w = tl.dot(b_A, b_kb)
+        tl.store(p_w, b_w.to(p_w.dtype.element_ty), boundary_check=(0, 1))
 
 
-def chunk_scaled_dot_kkt_fwd(
+def recompute_w_u_fwd(
     k: torch.Tensor,
+    v: torch.Tensor,
     beta: torch.Tensor,
-    g_cumsum: torch.Tensor | None = None,
-    cu_seqlens: torch.LongTensor | None = None,
-    chunk_size: int = 64,
-    output_dtype: torch.dtype = torch.float32,
-) -> torch.Tensor:
-    r"""
-    Compute beta * K * K^T.
+    g_cumsum: torch.Tensor,
+    A: torch.Tensor,
+    cu_seqlens: torch.LongTensor | None,
+) -> tuple[torch.Tensor, torch.Tensor]:
+    B, T, Hg, K, V = *k.shape, v.shape[-1]
+    H = v.shape[-2]
+    BT = A.shape[-1]
 
-    Args:
-        k (torch.Tensor):
-            The key tensor of shape `[B, T, H, K]`.
-        beta (torch.Tensor):
-            The beta tensor of shape `[B, T, H]`.
-        g_cumsum (torch.Tensor):
-            The cumulative sum of the gate tensor of shape `[B, T, H]`.
-            Default: None
-        cu_seqlens (torch.LongTensor):
-            The cumulative sequence lengths of the input tensor.
-            Default: None
-        chunk_size (int):
-            The chunk size. Default: 64.
-        output_dtype (torch.dtype):
-            The dtype of the output tensor. Default: `torch.float32`
-
-    Returns:
-        beta * K * K^T of shape `[B, T, H, BT]` where `BT` is the chunk size.
-    """
-
-    B, T, Hg, K = k.shape
-
-    H = beta.shape[-1]
-    BT = chunk_size
     chunk_indices = (
         prepare_chunk_indices(cu_seqlens, BT) if cu_seqlens is not None else None
     )
     NT = triton.cdiv(T, BT) if cu_seqlens is None else len(chunk_indices)
-    A = torch.empty(B, T, H, BT, device=k.device, dtype=output_dtype)
-    chunk_scaled_dot_kkt_fwd_kernel[(NT, B * H)](
+    BK = 64
+    BV = 64
+    u = torch.empty_like(v)
+    w = k.new_empty(B, T, H, K)
+    recompute_w_u_fwd_kernel[(NT, B * H)](
         k=k,
+        v=v,
         beta=beta,
-        g_cumsum=g_cumsum,
+        w=w,
+        u=u,
         A=A,
+        g=g_cumsum,
         cu_seqlens=cu_seqlens,
         chunk_indices=chunk_indices,
         T=T,
         H=H,
         Hg=Hg,
         K=K,
+        V=V,
         BT=BT,
-        BK=64,
-        num_warps=8,
+        BK=BK,
+        BV=BV,
+        num_warps=4,
         num_stages=3,
     )
-    return A
+    return w, u
+
+
+fwd_recompute_w_u = recompute_w_u_fwd
